pyproject.toml
[tool.poetry]
name = "modul-11"
version = "0.1.0"
description = ""
authors = ["Invil124 <113246425+Invil124@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "modul_11"}]

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.3"
fastapi = "^0.91.0"
alembic = "^1.9.3"
psycopg2 = "^2.9.5"
uvicorn = "^0.20.0"
pydantic = {extras = ["dotenv"], version = "^1.10.5"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
fastapi-mail = "^1.2.6"
redis = {extras = ["asyncio"], version = "^4.5.1"}
fastapi-limiter = "^0.1.5"
asyncio = "^3.4.3"
cloudinary = "^1.32.0"
pytest = "^7.2.2"
httpx = "^0.23.3"


[tool.poetry.group.dev.dependencies]
sphinx = "^6.1.3"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
